Moving testing back to the main directory and adding a few more tests in.
