--- src/DataLib/fendl32B_retrofit/tests/test_reaction_data.py
@@ -1,19 +1,0 @@
-
-# Import packages
-import pytest
-import reaction_data as rxd
-
-@pytest.mark.parametrize(
-    "particle, emitted_particle_string, exp",
-    [
-        ('n', 'np', 1),
-        ('p', 'np', 1),
-        ('t', 'np', 0),
-        ('n', '357n14d', 357),
-        ('p', '357n14d', 0),
-        ('d', '357n14d', 14),
-    ]
-)
-def test_count_emitted_particles(particle, emitted_particle_string, exp):
-    obs = rxd.count_emitted_particles(particle, emitted_particle_string)
-    assert obs == exp
